+++ src/index.ts
@@ -18,15 +18,15 @@ export * from './Components/Spinners/SparkleSpinner'
 export * from './Components/Spinners/SpinnerContainer'
 export * from './Components/Spinners/TechnoSpinner'
 export * from './Components/Spinners/WaitingText'
-export * from './Components/FormLayout/Form'
-export * from './Components/FormLayout/FormDivider'
-export * from './Components/FormLayout/FormElement'
-export * from './Components/FormLayout/FormElementError'
-export * from './Components/FormLayout/FormFooter'
-export * from './Components/FormLayout/FormHelpText'
-export * from './Components/FormLayout/FormLabel'
-export * from './Components/FormLayout/FormBox'
-export * from './Components/FormLayout/FormValidationElement'
+export * from './Components/Form/Form'
+export * from './Components/Form/FormDivider'
+export * from './Components/Form/FormElement'
+export * from './Components/Form/FormElementError'
+export * from './Components/Form/FormFooter'
+export * from './Components/Form/FormHelpText'
+export * from './Components/Form/FormLabel'
+export * from './Components/Form/FormBox'
+export * from './Components/Form/FormValidationElement'
 export * from './Components/ColorPicker/ColorPicker'
 export * from './Components/ColorPicker/ColorPickerSwatch'
 export * from './Components/EmptyState/EmptyState'
